move result display out as a component
